Refactor: Split codebase into modular components
This refactoring improves code organization and maintainability by:
- Using main.py as the entry point
- Splitting functionality into dedicated modules:
  - tiktok_scraper.py: Core TikTok interaction logic
  - browser_manager.py: Browser setup and management
  - airtable_manager.py: Airtable integration
  - file_handlers.py: File system operations
- Removing the old monolithic file

This modular approach makes the code easier to understand,
maintain, and modify. It also improves compatibility with
AI development tools like Windsurf by keeping files focused
and concise.
